remove: mkint() update: update IMDb during images update fix: IMDb rating not showing on browse page fix: IMDb block not showing genres fix: find_images() not caching remove: star.gif, looks terrible fix: formatting rules.php fix: tabs users.php

+++ include/html_functions.php
@@ -241,8 +241,12 @@ function write_css($data, $template)
 ";
         }
     }
-    file_put_contents(ROOT_DIR . "chat/css/{$template}/classcolors.css", $classdata . PHP_EOL);
-    file_put_contents(ROOT_DIR . "templates/{$template}/css/classcolors.css", $classdata . PHP_EOL);
+    if (file_exists(ROOT_DIR . "chat/css/{$template}")) {
+        file_put_contents(ROOT_DIR . "chat/css/{$template}/classcolors.css", $classdata . PHP_EOL);
+    }
+    if (file_exists(ROOT_DIR . "templates/{$template}/css")) {
+        file_put_contents(ROOT_DIR . "templates/{$template}/css/classcolors.css", $classdata . PHP_EOL);
+    }
 }
 
 /**
